app/requirements.txt
dash
plotly
requests
pandas
psycopg2-binary